+++ bootstrap/console/start.php
@@ -6,8 +6,8 @@
  */
 use RDev\Console\Commands\CommandCollection;
 use RDev\Console\Commands\Compilers\ICompiler;
-use RDev\Console\Kernels\Kernel;
 use RDev\Console\Requests\Parsers\IParser;
+use RDev\Framework\Console\Kernel;
 
 require_once __DIR__ . "/../start.php";