+++ configs/http/bootstrappers.php
@@ -28,6 +28,6 @@ return [
      *
      * List any HTTP bootstrappers you'd like here
      */
-    "Project\\Bootstrappers\\HTTP\\Routing\\Router",
-    "Project\\Bootstrappers\\HTTP\\Views\\Builders"
+    "Project\\Bootstrappers\\HTTP\\Views\\Builders",
+    "Project\\Bootstrappers\\Routing\\Router"
 ];